Initial Commit
This is the intial commit of the Model A Wiki Website. This commit is the finished product of the DevMArketer YouTube Tutorials on Multi-Authorization and the Basic Blog built into one application. There is no branding on the website just yet. After this commit, we will focus on small tweeks and improvements before moving on to Branding the website.

For debugging purposes the 'main' view has a compenent commented out. This component just tells the developer which authentication he currently has: 'user' or 'admin' or 'both'.